formatted dates on all pages. Also created and fully styled edit page. Lastly, need to integrate prior blog post data into frontend, and functionality to handle PUT request in the backend

+++ public/js/profile.js
@@ -36,62 +36,52 @@ const makeNewPost = async (e) => {
   }
 };
 
-const deletePost = async () => {
-  let id = deleteBtn.dataset.postId;
-  let res = await fetch(`/profile/deletePost/${id}`, {
-    headers: {
-      "Content-type": "Application/json",
-    },
-    method: "DELETE",
-  });
-
-  if (res.ok) {
-    let deletion = await res.json();
-
-    if (deletion) {
-      window.location.replace("/profile");
-    }
-  }
-};
-
 const updatePost = (id) => {};
 
 newPostBtn.addEventListener("click", makeNewPost);
 
-deleteBtn.addEventListener("click", deletePost);
-
-const blogCards = document.querySelectorAll(".recent-posts");
-
 const handleBlogCard = async (event) => {
-  console.log(event.target);
   const target = event.target;
-  const dataAttr = target.dataset.postId;
+  const cardContainer = target.closest(".recent-posts");
+  const postId = cardContainer.id;
 
   if (target.id === "delete-post-btn") {
-    let deletePost = await fetch(`/profile/deletePost/${dataAttr}`, {
-      headers: {
-        "Content-type": "Application/json",
-      },
-      method: "DELETE",
-    });
+    const confirmDelete = confirm(
+      "This is an irreversible action, are you sure you want to delete?"
+    );
+    if (confirmDelete === true) {
+      const deletePost = await fetch(`/profile/deletePost/${postId}`, {
+        headers: {
+          "Content-type": "Application/json",
+        },
+        method: "DELETE",
+      });
 
-    if (deletePost.ok) {
-      const successDelete = await deletePost.json();
+      if (deletePost.ok) {
+        const successDelete = await deletePost.json();
 
-      if (successDelete) {
-        // window.location.replace("/profile");
+        if (successDelete) {
+          window.location.replace("/profile");
+        }
       }
+    } else {
+      return;
     }
   } else if (target.id === "edit-btn") {
-    const editPost = await fetch(`/profile/updatePost/${dataAttr}`);
-    if (editPost.ok) {
-      const update = await editPost.json();
-      if (update) {
-        // window.location.replace('/')
-      }
-    }
+    window.location.replace("/edit");
+
+    // const editPost = await fetch(`/profile/updatePost/${dataAttr}`);
+    // if (editPost.ok) {
+    //   const update = await editPost.json();
+    //   if (update) {
+    //   }
+    // }
+  } else {
+    window.location.replace(`/posts/${postId}`);
   }
 };
+const blogCards = document.querySelectorAll(".recent-posts");
+
 blogCards.forEach((card) => {
   card.addEventListener("click", handleBlogCard);
 });
